Add dist-upgrade for MSYS2 system upgrade
Because of Windows's file locking, the typical

    $ pacman -Syu

for a full system upgrade doesn't work so well. It is best to upgrade
the runtime alone, restart MSYS2, then upgrade remaining packages. The
original process for this [1] was

    $ pacman --needed -Sy bash pacman pacman-mirrors msys2-runtime
    # restart MSYS2
    $ pacman -Syu

Since 2015-06-30, MSYS2 has shipped with an update-core script [2] to
simplify the first step of the upgrade process, although the
documentation was not updated to reflect this until February 2016 [1].

Add a dist-upgrade command, named for the analogous apt-get command,
that upgrades the MSYS2 core, and use it to call update-core on the
basis that the MSYS2 maintainers know better what to do than I.

The user must still restart and subsequently upgrade remaining packages
manually. There is no nice way to get around this but at present
update-core will remind the user.

With the standalone update-core command, this addition is arguably less
valuable, perhaps even counterproductive. On the other hand, the
indirection is counter-intuitive and the separation bad for
discoverability.

[1] https://github.com/msys2/msys2.github.io/commit/dfc4af1df6a3270febb356ac4ffd3e04c61e5f40
[2] https://github.com/Alexpux/MSYS2-pacman/commit/5c8d12aa1f767b29549e63654b29ccc6f5c66bdf